refactor(l5): move maintenance-bootstrap from src/ into server/maintenance behind an L5 adapter (issue 864)
Moved with git mv:
  src/maintenance-bootstrap.ts -> server/maintenance/maintenance-bootstrap.ts

The server/ version reads no process.env. The six tuning values and the poll
cadence arrive as fields of its existing single options parameter; the runner
and the sweep stay the single owners of their own defaults, exactly as before.
server/config/maintenance.ts already parses every one of those keys, so no new
reader was needed and server/config.ts is untouched. server/main.ts changes by
one import path only: it imports composeMaintenanceHandlers and
MAINTENANCE_GRAPH_AUTH, neither of which took an env value.

src/maintenance-bootstrap.ts is now the M9 adapter (under 60 code lines, every
relative import naming a server/ path, type imports included). It preserves the
old call forms for src/index.ts:43 and src/maintenance-sweep.test.ts: it reads
the six OPEN_BRAIN_MAINTENANCE_* overrides at call time and folds them into the
options it forwards, and it keeps maintenanceQueueEnabled, an env predicate with
no server/ caller. src/maintenance-bootstrap.test.ts and
src/maintenance-sweep.test.ts stay at their src/ paths untouched per M1
EXCEPTION and exercise the moved code through the adapter.

Imports of the handlers, queue runner, sweep and dream modules now name the
server/maintenance/ siblings directly; still-in-src imports (embedding,
embedding-repair) stay relative per M3.

Closure count: 27 -> 26.

Refs issue 864.
